Rename multiexposurespec to multispec and change attribute from exposures to spec
