+++ src/stdatamodels/jwst/datamodels/__init__.py
@@ -38,9 +38,8 @@ from .mask import MaskModel
 from .mrsxartcorr import MirMrsXArtCorrModel
 from .multicombinedspec import MultiCombinedSpecModel, WFSSMultiCombinedSpecModel
 from .multiexposure import MultiExposureModel
-from .multiexposurespec import WFSSMultiExposureSpecModel
 from .multislit import MultiSlitModel
-from .multispec import MultiSpecModel, MRSMultiSpecModel, TSOMultiSpecModel
+from .multispec import MultiSpecModel, MRSMultiSpecModel, TSOMultiSpecModel, WFSSMultiSpecModel
 from .nirspec_flat import NirspecFlatModel, NirspecQuadFlatModel
 from .nrm import NRMModel
 from .outlierifuoutput import OutlierIFUOutputModel
@@ -254,7 +253,7 @@ __all__ = [
     "WfssBkgModel",
     "WFSSCombinedSpecModel",
     "WFSSMultiCombinedSpecModel",
-    "WFSSMultiExposureSpecModel",
+    "WFSSMultiSpecModel",
     "WFSSSpecModel",
 ]
