+++ src/stdatamodels/jwst/datamodels/schemas/wfss_multispec.schema.yaml
@@ -1,12 +1,12 @@
 %YAML 1.1
 ---
 $schema: "http://stsci.edu/schemas/asdf/asdf-schema-1.1.0"
-id: "http://stsci.edu/schemas/jwst_datamodel/wfss_multiexposurespec.schema"
+id: "http://stsci.edu/schemas/jwst_datamodel/wfss_multispec.schema"
 allOf:
 - $ref: core.schema
 - type: object
   properties:
-    exposures:
+    spec:
       type: array
       title: An array of WFSS specmodels
       description: An array of WFSS multispecmodels, one model per exposure.
